Streamline some test code
